Renamed IAM files in utils/prometheus to focus on IAM policy. Cleaned up unused code. Switched to using global namespace variable in setup_prometheus_server.py.
