--- tests/e2e/utils/prometheus/createIRSA_AMPIngest.py
@@ -1,64 +1,0 @@
-from e2e.utils.prometheus.IAMRolesUtils import *
-from e2e.utils.utils import get_aws_account_id
-from e2e.fixtures.cluster import (
-    associate_iam_oidc_provider,
-    get_oidc_provider,
-)
-
-# Global Variables
-SERVICE_ACCOUNT_NAMESPACE = "monitoring"
-SERVICE_ACCOUNT_AMP_INGEST_NAME = "amp-iamproxy-ingest-service-account"
-SERVICE_ACCOUNT_IAM_AMP_INGEST_ROLE = "amp-iamproxy-ingest-role"
-SERVICE_ACCOUNT_IAM_AMP_INGEST_POLICY = "AMPIngestPolicy"
-TRUST_POLICY_FILE_NAME = "AMPTrustPolicy"
-PERMISSION_POLICY_FILE_PATH = "../../deployments/add-ons/prometheus/AMPIngestPermissionPolicy"
-
-# Create a trust policy json file
-def create_AMP_trust_policy_file():
-    aws_account_id = get_aws_account_id()
-    oidc_provider = get_oidc_provider(CLUSTER_NAME, CLUSTER_REGION)
-    trust_policy = open(TRUST_POLICY_FILE_NAME + '.json', 'w')
-    trust_policy.write('\n'.join([
-        '{',
-        '  "Version": "2012-10-17",',
-        '  "Statement": [',
-        '    {',
-        '      "Effect": "Allow",',
-        '      "Principal": {',
-        f'        "Federated": "arn:aws:iam::{aws_account_id}:oidc-provider/{oidc_provider}"',
-        '      },',
-        '      "Action": "sts:AssumeRoleWithWebIdentity",',
-        '      "Condition": {',
-        '        "StringEquals": {',
-        f'          "{oidc_provider}:sub": "system:serviceaccount:{SERVICE_ACCOUNT_NAMESPACE}:{SERVICE_ACCOUNT_AMP_INGEST_NAME}"',
-        '        }',
-        '      }',
-        '    }',
-        '  ]',
-        '}']))
-    trust_policy.close()
-
-def create_AMP_ingest_policy():
-    return create_policy_if_not_exist(SERVICE_ACCOUNT_IAM_AMP_INGEST_POLICY, PERMISSION_POLICY_FILE_PATH)
-
-#def setup_ingest_role(cluster_name, cluster_region):
-#    global CLUSTER_NAME, CLUSTER_REGION
-#    CLUSTER_NAME = cluster_name
-#    CLUSTER_REGION = cluster_region
-#    
-#    setup_role_and_policy(
-#        SERVICE_ACCOUNT_IAM_AMP_INGEST_ROLE, SERVICE_ACCOUNT_IAM_AMP_INGEST_POLICY,
-#        TRUST_POLICY_FILE_NAME, PERMISSION_POLICY_FILE_PATH,
-#        create_AMP_trust_policy_file)
-#    
-#    associate_iam_oidc_provider(CLUSTER_NAME, CLUSTER_REGION)
-
-#def delete_ingest_role():
-#    print("Deleting policy function called.")
-#    detach_IAM_policy(SERVICE_ACCOUNT_IAM_AMP_INGEST_ROLE, SERVICE_ACCOUNT_IAM_AMP_INGEST_POLICY)
-#    delete_IAM_policy(SERVICE_ACCOUNT_IAM_AMP_INGEST_POLICY)
-#    print("Deleting role function called.")
-#    delete_IAM_role(SERVICE_ACCOUNT_IAM_AMP_INGEST_ROLE)
-
-def delete_policy():
-    delete_IAM_policy(SERVICE_ACCOUNT_IAM_AMP_INGEST_POLICY)
