+++ tests/e2e/utils/prometheus/setup_prometheus_server.py
@@ -3,13 +3,7 @@ import time
 import os
 import boto3
 import tempfile
-from e2e.utils.prometheus.createIRSA_AMPIngest import (
-#    setup_ingest_role,
-#    delete_ingest_role,
-    create_AMP_ingest_policy,
-    delete_policy,
-)
-from e2e.utils.utils import get_aws_account_id
+from e2e.utils.prometheus.createPolicy_AMPIngest import *
 from e2e.utils.utils import wait_for
 from e2e.fixtures.cluster import associate_iam_oidc_provider
 
@@ -30,12 +24,6 @@ def make_new_file_with_replacement(file_path, original_to_replacement_dict):
         new_file.write(new_file_contents)
     return new_file_path
 
-def update_AMP_service_account_id(AMP_service_account_file_path):
-    aws_account_id = get_aws_account_id()
-    original_to_replacement_dict = {}
-    original_to_replacement_dict['<my-account-id>'] = aws_account_id
-    return make_new_file_with_replacement(AMP_service_account_file_path, original_to_replacement_dict)
-    
 def update_config_map_AMP_workspace(workspace_id, region, config_map_file_path):
     original_to_replacement_dict = {}
     original_to_replacement_dict['<my-workspace-id>'] = workspace_id
@@ -57,27 +45,21 @@ def create_AMP_workspace(region):
     
     return workspace_id
 
-def set_up_prometheus_for_AMP(cluster_name, region):
-    workspace_id = create_AMP_workspace(region)
-    
-#    setup_ingest_role(cluster_name, region)
-
-#    # Edit AMP service account to use account-id
-#    new_AMP_service_account = update_AMP_service_account_id(f'{prometheus_yaml_files_directory}/AMP-service-account.yaml')
-
-    print("About to associate IAM with OIDC provider.")
+def set_up_prometheus_for_AMP(cluster_name, region):    
     associate_iam_oidc_provider(cluster_name, region)
-    print("Finished associating IAM with OIDC provider.")
 
-    # Create AMP service account:
     policy_arn = create_AMP_ingest_policy()
-    create_AMP_service_account_command = f'eksctl create iamserviceaccount --name amp-iamproxy-ingest-service-account --role-name amp-iamproxy-ingest-role --namespace monitoring --cluster {cluster_name} --attach-policy-arn {policy_arn} --override-existing-serviceaccounts --approve --region {region}'.split()
+    
+    # Create AMP service account:
+    create_AMP_service_account_command = f'eksctl create iamserviceaccount --name {SERVICE_ACCOUNT_AMP_INGEST_NAME} --role-name {SERVICE_ACCOUNT_IAM_AMP_INGEST_ROLE} --namespace {PROMETHEUS_NAMESPACE} --cluster {cluster_name} --attach-policy-arn {policy_arn} --override-existing-serviceaccounts --approve --region {region}'.split()
     subprocess.call(create_AMP_service_account_command, encoding="utf-8")
 
+    workspace_id = create_AMP_workspace(region)
+    
     # Edit config map to use workspace-id and region
     new_config_map = update_config_map_AMP_workspace(workspace_id, region, f'{prometheus_yaml_files_directory}/config-map.yaml')
     
-    create_namespace_command = 'kubectl create namespace monitoring'.split()
+    create_namespace_command = f'kubectl create namespace {PROMETHEUS_NAMESPACE}'.split()
     subprocess.call(create_namespace_command)
     
     kustomize_build_command = f'kustomize build {prometheus_yaml_files_directory}'.split()
@@ -98,9 +80,10 @@ def set_up_prometheus_for_AMP(cluster_name, region):
 def set_up_prometheus_port_forwarding():
     print("Entered prometheus port_forwarding_function.")
     def get_prometheus_pod_name():
-        get_pod_name_command = 'kubectl get pods --namespace=monitoring'.split()
+        get_pod_name_command = f'kubectl get pods --namespace={PROMETHEUS_NAMESPACE}'.split()
+        print(' '.join(get_pod_name_command))
         pod_list = subprocess.check_output(get_pod_name_command, encoding="utf-8").split('\n')[1:]
-        print(f"Monitoring Pods list: {pod_list}")
+        print(f"{PROMETHEUS_NAMESPACE} Pods list: {pod_list}")
         assert 0 < len(pod_list)
         prometheus_pod_name = None
         for pod in pod_list:
@@ -110,12 +93,12 @@ def set_up_prometheus_port_forwarding():
             if (('p' == pod_name[0])
                 and ('prometheus' == pod_name_array[0])
                 and ('deployment' == pod_name_array[1])):
-                describe_pod_command = f'kubectl describe pod {pod_name} --namespace monitoring'.split()
+                describe_pod_command = f'kubectl describe pod {pod_name} --namespace {PROMETHEUS_NAMESPACE}'.split()
                 print('\n\nAbout to call describe pod command!\n\n')
                 print(subprocess.check_output(describe_pod_command, encoding="utf-8"))
                 print('\n\nDescribe Finished\n\n')
-                get_ns_pod_logs_command = f'kubectl logs {pod_name} --namespace monitoring'.split()
-                print("\n\nAbout to get monitoring namespace logs!\n\n")
+                get_ns_pod_logs_command = f'kubectl logs {pod_name} --namespace {PROMETHEUS_NAMESPACE}'.split()
+                print(f"\n\nAbout to get {PROMETHEUS_NAMESPACE}/{pod_name} logs!\n\n")
                 print(subprocess.check_output(get_ns_pod_logs_command, encoding="utf-8"))
                 print('\n\nLogs Finished\n\n')
                 pod_status = pod.split()[1]
@@ -130,7 +113,7 @@ def set_up_prometheus_port_forwarding():
     if None == prometheus_pod_name:
         raise ValueError("Prometheus Pod Not Running.")
 
-    set_up_port_forwarding_command = f'kubectl port-forward {prometheus_pod_name} 9090:9090 -n monitoring'.split()
+    set_up_port_forwarding_command = f'kubectl port-forward {prometheus_pod_name} 9090:9090 -n {PROMETHEUS_NAMESPACE}'.split()
     print(" ".join(set_up_port_forwarding_command))
     port_forwarding_process = subprocess.Popen(set_up_port_forwarding_command)
     time.sleep(10)  # Wait 10 seconds for port forwarding to open
@@ -170,12 +153,8 @@ def check_AMP_connects_to_prometheus(region, workspace_id, expected_value):
     return AMP_create_experiment_count
 
 def delete_AMP_resources(cluster_name, region, workspace_id):
-#    print("About to delete ingest role.")
-#    delete_ingest_role()
-#    print("Finished deleting ingest role.")
-
     print("About to delete service account and role.")
-    delete_AMP_service_account_command = f'eksctl delete iamserviceaccount --name amp-iamproxy-ingest-service-account --namespace monitoring --cluster {cluster_name} --region {region}'.split()
+    delete_AMP_service_account_command = f'eksctl delete iamserviceaccount --name amp-iamproxy-ingest-service-account --namespace {PROMETHEUS_NAMESPACE} --cluster {cluster_name} --region {region}'.split()
     subprocess.call(delete_AMP_service_account_command, encoding="utf-8")
     print("Finished deleting service account and role.")
